Move Batching into a Schema plugin
